Wrapper Long

+++ src/main/java/edu/icet/repository/NotificationRepository.java
@@ -3,5 +3,5 @@ package edu.icet.repository;
 import edu.icet.entity.NotificationEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NotificationRepository extends JpaRepository<NotificationEntity,Integer> {
+public interface NotificationRepository extends JpaRepository<NotificationEntity,Long> {
 }
